feat: 메인 페이지에 필요한 API Response 받아오기 :

+++ src/main/java/com/codesquad_team01/issue_tracker/issue/dto/response/IssueListResponse.java
@@ -7,9 +7,9 @@ public record IssueListResponse(
         List<IssueResponse> issues
 ) {
     public record Metadata(
-            int openIssueCount,
-            int closedIssueCount,
-            int labelCount,
-            int milestoneCount
+            long openIssueCount,
+            long closedIssueCount,
+            long labelCount,
+            long milestoneCount
     ) {}
 }
